+++ modules/domain/src/main/java/com/eastbarnetschool/ordermatchingengine/domain/PriceLevel.java
@@ -3,16 +3,16 @@ package com.eastbarnetschool.ordermatchingengine.domain;
 import java.util.LinkedList;
 
 public class PriceLevel {
-    private Long price;
+    private Double price;
     private LinkedList<Order> orders;
 
-    public PriceLevel(Long price, Order order) {
+    public PriceLevel(Double price, Order order) {
         this.price = price;
         this.orders = new LinkedList<>();
         orders.add(order);
     }
 
-    public Long getPrice() {
+    public Double getPrice() {
         return price;
     }
